fix: cleanup (#157)
* fix-csp

* cleanup

* Update src/lib/e2esdk.ts


* fix: use @socialgouv/e2esdk-crypto.fileMetadataSchema

* fix: readAndEncryptFile signature

* feat: use modal for file previews

* dummy

---------
