+++ package.json
@@ -27,37 +27,37 @@
     "keycloak": "chmod +x keycloak/scripts/keycloak.sh && keycloak/scripts/keycloak.sh"
   },
   "dependencies": {
-    "@codegouvfr/react-dsfr": "^0.48.1",
-    "@emotion/react": "^11.10.6",
-    "@emotion/server": "^11.10.0",
-    "@emotion/styled": "^11.10.6",
+    "@codegouvfr/react-dsfr": "^0.58.1",
+    "@emotion/react": "^11.11.0",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.11.1",
-    "@mui/x-data-grid": "^6.2.1",
+    "@mui/material": "^5.13.3",
+    "@mui/x-data-grid": "^6.5.0",
     "@mui/x-date-pickers": "^5.0.11",
     "@sentry/nextjs": "^6.17.9",
     "@socialgouv/e2esdk-client": "1.0.0-beta.26",
     "@socialgouv/e2esdk-crypto": "1.0.0-beta.18",
     "@socialgouv/e2esdk-devtools": "1.0.0-beta.35",
     "@socialgouv/e2esdk-react": "1.0.0-beta.25",
-    "@socialgouv/matomo-next": "^1.4.0",
+    "@socialgouv/matomo-next": "^1.6.1",
     "boring-avatars": "^1.7.0",
     "dayjs": "^1.11.7",
-    "formidable": "v3",
+    "formidable": "3.2.5",
     "js-image-generator": "^1.0.4",
     "next": "13.1.0",
-    "next-auth": "^4.19.1",
+    "next-auth": "^4.22.1",
     "next-transpile-modules": "^10.0.0",
     "random-words": "^1.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
-    "react-hook-form": "^7.43.9",
+    "react-hook-form": "^7.44.2",
     "tss-react": "^4.6.0",
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
+    "@babel/core": "^7.22.1",
     "@storybook/addon-actions": "^6.4.19",
     "@storybook/addon-docs": "^6.5.15",
     "@storybook/addon-essentials": "^6.4.19",
@@ -72,20 +72,20 @@
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
     "cypress": "^9.4.1",
-    "eslint": "8.30.0",
-    "eslint-config-next": "13.1.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-storybook": "^0.5.7",
+    "eslint": "8.41.0",
+    "eslint-config-next": "13.4.4",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-storybook": "^0.6.12",
     "gh-pages": "^4.0.0",
     "hasura-cli": "^2.25.0",
-    "husky": "^8.0.1",
+    "husky": "^8.0.3",
     "is-ci": "^3.0.1",
     "jest": "^27.5.1",
     "lint-staged": "^12.3.7",
-    "node-talisman": "^1.28.0",
+    "node-talisman": "^1.29.10",
     "start-server-and-test": "^1.14.0",
     "storybook-dark-mode": "^2.0.5",
-    "tsx": "^3.12.2",
+    "tsx": "^3.12.7",
     "typescript": "4.9.4"
   }
 }
